fix: Ensure it actually builds with prerender mode
