+++ packages/ember-build-prerender/lib/commands/prerender.js
@@ -109,6 +109,8 @@ module.exports = {
 
     let Builder = this.project.require('ember-cli/lib/models/builder');
 
+    process.env.PRERENDER = 'true';
+
     let builder = new Builder({
       ui: this.ui,
       outputPath,
