+++ packages/test-app/node-tests/fixtures/build-dist/assets/test-app-2cf0aa3042605f84142e2c723b702f5f.js
@@ -1,0 +1,58 @@
+"use strict"
+define("test-app/app",["exports","@ember/application","ember-resolver","ember-load-initializers","test-app/config/environment"],(function(e,t,n,r,l){function a(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}Object.defineProperty(e,"__esModule",{value:!0}),e.default=void 0
+class i extends t.default{constructor(){super(...arguments),a(this,"modulePrefix",l.default.modulePrefix),a(this,"podModulePrefix",l.default.podModulePrefix),a(this,"Resolver",n.default)}}e.default=i,(0,r.default)(i,l.default.modulePrefix)})),define("test-app/application/route",["exports","@ember/routing/route"],(function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=void 0
+class n extends t.default{}e.default=n})),define("test-app/application/template",["exports","@ember/template-factory"],(function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=void 0
+var n=(0,t.createTemplateFactory)({id:"dSv1f+C5",block:'[[[1,[28,[35,0],["Prerender Test App"],null]],[1,"\\n\\n"],[10,"h1"],[12],[1,"\\n  Welcome to the Prerender Test App\\n"],[13],[1,"\\n\\n"],[46,[28,[37,2],null,null],null,null,null]],[],false,["page-title","component","-outlet"]]',moduleName:"test-app/application/template.hbs",isStrictMode:!1})
+e.default=n})),define("test-app/component-managers/glimmer",["exports","@glimmer/component/-private/ember-component-manager"],(function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})})),define("test-app/components/button",["exports","@ember/component","@ember/template-factory","@glimmer/component","@ember/object"],(function(e,t,n,r,l){var a
+Object.defineProperty(e,"__esModule",{value:!0}),e.default=void 0
+const i=(0,n.createTemplateFactory)({id:"lBZrmgLv",block:'[[[11,"button"],[24,4,"button"],[4,[38,0],["click",[30,0,["showAlert"]]],null],[12],[1,"\\n  Show alert\\n"],[13]],[],false,["on"]]',moduleName:"test-app/components/button.hbs",isStrictMode:!1})
+let o=(a=class extends r.default{showAlert(){window.alert("Interactivity is working!")}},u=a.prototype,p="showAlert",d=[l.action],s=Object.getOwnPropertyDescriptor(a.prototype,"showAlert"),c=a.prototype,f={},Object.keys(s).forEach((function(e){f[e]=s[e]})),f.enumerable=!!f.enumerable,f.configurable=!!f.configurable,("value"in f||f.initializer)&&(f.writable=!0),f=d.slice().reverse().reduce((function(e,t){return t(u,p,e)||e}),f),c&&void 0!==f.initializer&&(f.value=f.initializer?f.initializer.call(c):void 0,f.initializer=void 0),void 0===f.initializer&&(Object.defineProperty(u,p,f),f=null),a)
+var u,p,d,s,c,f
+e.default=o,(0,t.setComponentTemplate)(i,o)})),define("test-app/components/is-prerendered",["exports","@ember/component","@ember/template-factory","@ember/service","@glimmer/component"],(function(e,t,n,r,l){var a,i
+Object.defineProperty(e,"__esModule",{value:!0}),e.default=void 0
+const o=(0,n.createTemplateFactory)({id:"vxj8ksRz",block:'[[[10,0],[12],[1,"\\n  Is prerendered:\\n  "],[10,1],[14,"data-test-is-prerendered",""],[12],[1,"\\n    "],[1,[30,0,["isPrerender"]]],[1,"\\n  "],[13],[1,"\\n"],[13]],[],false,[]]',moduleName:"test-app/components/is-prerendered.hbs",isStrictMode:!1})
+let u=(a=class extends l.default{constructor(){var e,t,n,r
+super(...arguments),e=this,t="prerender",r=this,(n=i)&&Object.defineProperty(e,t,{enumerable:n.enumerable,configurable:n.configurable,writable:n.writable,value:n.initializer?n.initializer.call(r):void 0})}get isPrerender(){return this.prerender.isPrerender}},p=a.prototype,d="prerender",s=[r.inject],c={configurable:!0,enumerable:!0,writable:!0,initializer:null},m={},Object.keys(c).forEach((function(e){m[e]=c[e]})),m.enumerable=!!m.enumerable,m.configurable=!!m.configurable,("value"in m||m.initializer)&&(m.writable=!0),m=s.slice().reverse().reduce((function(e,t){return t(p,d,e)||e}),m),f&&void 0!==m.initializer&&(m.value=m.initializer?m.initializer.call(f):void 0,m.initializer=void 0),void 0===m.initializer&&(Object.defineProperty(p,d,m),m=null),i=m,a)
+var p,d,s,c,f,m
+e.default=u,(0,t.setComponentTemplate)(o,u)})),define("test-app/components/lorem-ipsum",["exports","@ember/component","@ember/template-factory","@glimmer/component"],(function(e,t,n,r){Object.defineProperty(e,"__esModule",{value:!0}),e.default=void 0
+const l=(0,n.createTemplateFactory)({id:"5G9TbalM",block:'[[[42,[28,[37,1],[[28,[37,1],[[30,0,["items"]]],null]],null],null,[[[1,"  "],[10,2],[12],[1,"\\n    Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod\\n    tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam,\\n    quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo\\n    consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse\\n    cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat\\n    non proident, sunt in culpa qui officia deserunt mollit anim id est\\n    laborum.\\n  "],[13],[1,"\\n"]],[]],null]],[],false,["each","-track-array"]]',moduleName:"test-app/components/lorem-ipsum.hbs",isStrictMode:!1})
+class a extends r.default{get items(){let e=this.args.count||1,t=[]
+for(let n=0;n<e;n++)t.push(n)
+return t}}e.default=a,(0,t.setComponentTemplate)(l,a)})),define("test-app/helpers/app-version",["exports","@ember/component/helper","test-app/config/environment","ember-cli-app-version/utils/regexp"],(function(e,t,n,r){function l(e){let t=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{}
+const l=n.default.APP.version
+let a=t.versionOnly||t.hideSha,i=t.shaOnly||t.hideVersion,o=null
+return a&&(t.showExtended&&(o=l.match(r.versionExtendedRegExp)),o||(o=l.match(r.versionRegExp))),i&&(o=l.match(r.shaRegExp)),o?o[0]:l}Object.defineProperty(e,"__esModule",{value:!0}),e.appVersion=l,e.default=void 0
+var a=(0,t.helper)(l)
+e.default=a})),define("test-app/helpers/page-title",["exports","ember-page-title/helpers/page-title"],(function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=void 0
+var n=t.default
+e.default=n})),define("test-app/index/template",["exports","@ember/template-factory"],(function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=void 0
+var n=(0,t.createTemplateFactory)({id:"5CX0GXuK",block:'[[[10,"h2"],[12],[1,"\\n  Index page\\n"],[13],[1,"\\n\\n"],[10,"ul"],[12],[1,"\\n  "],[10,"li"],[12],[1,"\\n    "],[8,[39,0],null,[["@route"],["index"]],[["default"],[[[[1,"Index page"]],[]]]]],[1,"\\n  "],[13],[1,"\\n  "],[10,"li"],[12],[1,"\\n    "],[8,[39,0],null,[["@route"],["page-1"]],[["default"],[[[[1,"Page 1"]],[]]]]],[1,"\\n  "],[13],[1,"\\n  "],[10,"li"],[12],[1,"\\n    "],[8,[39,0],null,[["@route"],["page-2.index"]],[["default"],[[[[1,"Page 2"]],[]]]]],[1,"\\n  "],[13],[1,"\\n  "],[10,"li"],[12],[1,"\\n    "],[8,[39,0],null,[["@route"],["page-2.sub-page"]],[["default"],[[[[1,"Sub page of Page 2"]],[]]]]],[1,"\\n  "],[13],[1,"\\n"],[42,[28,[37,2],[[28,[37,2],[[28,[37,3],[1,2,3,4,5],null]],null]],null],null,[[[1,"    "],[10,"li"],[12],[1,"\\n      "],[8,[39,0],null,[["@route","@model"],["numbers.show",[30,1]]],[["default"],[[[[1,"\\n        Number page\\n        "],[1,[30,1]],[1,"\\n      "]],[]]]]],[1,"\\n    "],[13],[1,"\\n"]],[1]],null],[13],[1,"\\n\\n"],[10,0],[12],[1,"\\n  "],[10,"img"],[14,"src","/assets/vienna-9891f9bec17a731c557125a4a3a2f7a4.jpg"],[14,"alt","Vienna"],[12],[13],[1,"\\n"],[13],[1,"\\n\\n"],[10,0],[12],[1,"\\n  "],[8,[39,4],null,null,null],[1,"\\n"],[13],[1,"\\n\\n"],[8,[39,5],null,null,null]],["num"],false,["link-to","each","-track-array","array","button","is-prerendered"]]',moduleName:"test-app/index/template.hbs",isStrictMode:!1})
+e.default=n})),define("test-app/initializers/app-version",["exports","ember-cli-app-version/initializer-factory","test-app/config/environment"],(function(e,t,n){let r,l
+Object.defineProperty(e,"__esModule",{value:!0}),e.default=void 0,n.default.APP&&(r=n.default.APP.name,l=n.default.APP.version)
+var a={name:"App Version",initialize:(0,t.default)(r,l)}
+e.default=a})),define("test-app/initializers/container-debug-adapter",["exports","ember-resolver/resolvers/classic/container-debug-adapter"],(function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=void 0
+var n={name:"container-debug-adapter",initialize(){(arguments[1]||arguments[0]).register("container-debug-adapter:main",t.default)}}
+e.default=n})),define("test-app/initializers/export-application-global",["exports","ember","test-app/config/environment"],(function(e,t,n){function r(){var e=arguments[1]||arguments[0]
+if(!1!==n.default.exportApplicationGlobal){var r
+if("undefined"!=typeof window)r=window
+else if("undefined"!=typeof global)r=global
+else{if("undefined"==typeof self)return
+r=self}var l,a=n.default.exportApplicationGlobal
+l="string"==typeof a?a:t.default.String.classify(n.default.modulePrefix),r[l]||(r[l]=e,e.reopen({willDestroy:function(){this._super.apply(this,arguments),delete r[l]}}))}}Object.defineProperty(e,"__esModule",{value:!0}),e.default=void 0,e.initialize=r
+var l={name:"export-application-global",initialize:r}
+e.default=l})),define("test-app/instance-initializers/prerender",["exports","ember-build-prerender/instance-initializers/prerender"],(function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})})),define("test-app/instance-initializers/setup-rehydrate",["exports","ember-build-prerender/instance-initializers/setup-rehydrate"],(function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})})),define("test-app/numbers/show/route",["exports","@ember/routing/route"],(function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=void 0
+class n extends t.default{model(e){return e.number}}e.default=n})),define("test-app/numbers/show/template",["exports","@ember/template-factory"],(function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=void 0
+var n=(0,t.createTemplateFactory)({id:"c0daKBXp",block:'[[[10,"h3"],[12],[1,"Number "],[1,[30,0,["model"]]],[13],[1,"\\n\\n"],[10,2],[12],[1,"\\n  "],[8,[39,0],null,[["@route"],["index"]],[["default"],[[[[1,"Back to index"]],[]]]]],[1,"\\n"],[13]],[],false,["link-to"]]',moduleName:"test-app/numbers/show/template.hbs",isStrictMode:!1})
+e.default=n})),define("test-app/numbers/template",["exports","@ember/template-factory"],(function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=void 0
+var n=(0,t.createTemplateFactory)({id:"pDIqKtOM",block:'[[[10,"h2"],[12],[1,"Numbers"],[13],[1,"\\n\\n"],[46,[28,[37,1],null,null],null,null,null]],[],false,["component","-outlet"]]',moduleName:"test-app/numbers/template.hbs",isStrictMode:!1})
+e.default=n})),define("test-app/page-1/template",["exports","@ember/template-factory"],(function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=void 0
+var n=(0,t.createTemplateFactory)({id:"NEhI40Bo",block:'[[[10,"h2"],[12],[1,"Page 1"],[13],[1,"\\n\\n"],[8,[39,0],null,[["@count"],[2]],null],[1,"\\n\\n"],[10,2],[12],[1,"\\n  "],[8,[39,1],null,[["@route"],["index"]],[["default"],[[[[1,"Back to index"]],[]]]]],[1,"\\n"],[13]],[],false,["lorem-ipsum","link-to"]]',moduleName:"test-app/page-1/template.hbs",isStrictMode:!1})
+e.default=n})),define("test-app/page-2/index/template",["exports","@ember/template-factory"],(function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=void 0
+var n=(0,t.createTemplateFactory)({id:"WCFHeIgh",block:'[[[10,"h3"],[12],[1,"Index"],[13],[1,"\\n\\n"],[8,[39,0],null,[["@count"],[1]],null],[1,"\\n\\n"],[10,2],[12],[1,"\\n  "],[8,[39,1],null,[["@route"],["index"]],[["default"],[[[[1,"Back to index"]],[]]]]],[1,"\\n"],[13]],[],false,["lorem-ipsum","link-to"]]',moduleName:"test-app/page-2/index/template.hbs",isStrictMode:!1})
+e.default=n})),define("test-app/page-2/sub-page/template",["exports","@ember/template-factory"],(function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=void 0
+var n=(0,t.createTemplateFactory)({id:"YFXr+hE5",block:'[[[10,"h3"],[12],[1,"Sub page"],[13],[1,"\\n\\n"],[8,[39,0],null,[["@count"],[3]],null],[1,"\\n\\n"],[10,2],[12],[1,"\\n  "],[8,[39,1],null,[["@route"],["index"]],[["default"],[[[[1,"Back to index"]],[]]]]],[1,"\\n"],[13]],[],false,["lorem-ipsum","link-to"]]',moduleName:"test-app/page-2/sub-page/template.hbs",isStrictMode:!1})
+e.default=n})),define("test-app/page-2/template",["exports","@ember/template-factory"],(function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),e.default=void 0
+var n=(0,t.createTemplateFactory)({id:"ONOmJetY",block:'[[[10,"h2"],[12],[1,"Page 2"],[13],[1,"\\n\\n"],[46,[28,[37,1],null,null],null,null,null]],[],false,["component","-outlet"]]',moduleName:"test-app/page-2/template.hbs",isStrictMode:!1})
+e.default=n})),define("test-app/router",["exports","@ember/routing/router","test-app/config/environment"],(function(e,t,n){function r(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}Object.defineProperty(e,"__esModule",{value:!0}),e.default=void 0
+class l extends t.default{constructor(){super(...arguments),r(this,"location",n.default.locationType),r(this,"rootURL",n.default.rootURL)}}e.default=l,l.map((function(){this.route("page-1"),this.route("page-2",(function(){this.route("sub-page")})),this.route("numbers",(function(){this.route("show",{path:"/:number"})}))}))})),define("test-app/services/page-title-list",["exports","ember-page-title/services/page-title-list"],(function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})})),define("test-app/services/page-title",["exports","ember-page-title/services/page-title"],(function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})})),define("test-app/services/prerender",["exports","ember-build-prerender/services/prerender"],(function(e,t){Object.defineProperty(e,"__esModule",{value:!0}),Object.defineProperty(e,"default",{enumerable:!0,get:function(){return t.default}})})),define("test-app/config/environment",[],(function(){try{var e="test-app/config/environment",t=document.querySelector('meta[name="'+e+'"]').getAttribute("content"),n={default:JSON.parse(decodeURIComponent(t))}
+return Object.defineProperty(n,"__esModule",{value:!0}),n}catch(r){throw new Error('Could not read config from meta tag with name "'+e+'".')}})),runningTests||require("test-app/app").default.create({name:"test-app",version:"0.1.2+ed61cfd6"})
