+++ packages/test-app/node-tests/fixtures/expected/_empty.html
@@ -7,7 +7,8 @@
     <meta name="description" content="" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     
-<meta name="test-app/config/environment" content="%7B%22modulePrefix%22%3A%22test-app%22%2C%22environment%22%3A%22production%22%2C%22rootURL%22%3A%22%2F%22%2C%22locationType%22%3A%22auto%22%2C%22EmberENV%22%3A%7B%22FEATURES%22%3A%7B%7D%2C%22EXTEND_PROTOTYPES%22%3A%7B%22Date%22%3Afalse%7D%2C%22_APPLICATION_TEMPLATE_WRAPPER%22%3Afalse%2C%22_DEFAULT_ASYNC_OBSERVERS%22%3Atrue%2C%22_JQUERY_INTEGRATION%22%3Afalse%2C%22_TEMPLATE_ONLY_GLIMMER_COMPONENTS%22%3Atrue%7D%2C%22APP%22%3A%7B%22name%22%3A%22test-app%22%2C%22version%22%3A%220.1.1%2B7a667a98%22%7D%2C%22exportApplicationGlobal%22%3Afalse%7D" />
+<meta name="test-app/config/environment" content="%7B%22modulePrefix%22%3A%22test-app%22%2C%22environment%22%3A%22production%22%2C%22rootURL%22%3A%22%2F%22%2C%22locationType%22%3A%22auto%22%2C%22EmberENV%22%3A%7B%22FEATURES%22%3A%7B%7D%2C%22EXTEND_PROTOTYPES%22%3A%7B%22Date%22%3Afalse%7D%2C%22_APPLICATION_TEMPLATE_WRAPPER%22%3Afalse%2C%22_DEFAULT_ASYNC_OBSERVERS%22%3Atrue%2C%22_JQUERY_INTEGRATION%22%3Afalse%2C%22_TEMPLATE_ONLY_GLIMMER_COMPONENTS%22%3Atrue%7D%2C%22APP%22%3A%7B%22name%22%3A%22test-app%22%2C%22version%22%3A%220.1.2%2Bed61cfd6%22%7D%2C%22exportApplicationGlobal%22%3Afalse%7D" />
+
 
     <link integrity="" rel="stylesheet" href="/assets/vendor-d41d8cd98f00b204e9800998ecf8427e.css" />
     <link integrity="" rel="stylesheet" href="/assets/test-app-98ee95eb6b847855c5a1fa354975f609.css" />
@@ -17,8 +18,8 @@
   <body>
     
 
-    <script src="/assets/vendor-8ff1b60bcef19b374ec2c73a8d031624.js" integrity="sha256-4zzMl8w7eCUJU/7fIKmydihc4U4hqAE5nlixGob+c80= sha512-FdxIGroldOWhxjO9L1MYVWmiP62/pNlRuRE3TYdIKI/9HIluDzJNeGO6ffCYZ5QUSm4XU2k5egqX97BX98fcBw==" ></script>
-    <script src="/assets/test-app-e6f4803c0665a8d7a44e35c5f1e821ee.js" integrity="sha256-aN86SewTVu2YargpkLUt2P//I6WSBV7+swcw+GUtsVE= sha512-N9Ss5HLqRj8i2nTnBID4BfZ8pGsNyHRgXMPvvjmltjYeJ0o7FIR3EZId9gHp/YCFL8PbhA/oglnrb+MwXObwNA==" ></script>
+    <script src="/assets/vendor-dea3385c49eb2ba18e00fe3fc076c618.js" integrity="sha256-g5C78S2jNsi8s8qJG6uw/vZ95SyysfKJ/vI36a6oZh4= sha512-ibjrtUde4Wmacz+1TxpVpQrrdY/cWiK2A6djXaZULEMOKc85HseT62r2eTbOrr5LsjTzyVEF55EQlfPDKfLOdg==" ></script>
+    <script src="/assets/test-app-2cf0aa3042605f84142e2c723b702f5f.js" integrity="sha256-+2v49sRkSueztB+ZYhx2EpIKmCx7ibuve47S/Yl46nM= sha512-fLR46btrDcjyxDZQ/au0a5mRg3ec2tGtWyJePGnyU8EuV48x7y3AZZDLr5Tifzqjl9BcCdM6s14sTuKgOex0cw==" ></script>
 
     
   </body>
